Improved character positioning.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@829 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Adventure/AppBuilder/AppBuilder/BlitObjects.cpp
@@ -299,7 +299,7 @@ void MirrorObject::renderCharacter(CharacterObject* chr){
     chr->updateState(true, false);
     Vec2i oldpos = chr->getPosition();
     Vec2i newpos = oldpos;
-    int mirrorbase = (mDepth+1)*Engine::instance()->getWalkGridSize(false);
+    int mirrorbase = (mDepth+1)*(int)Engine::instance()->getWalkGridSize(false);
     newpos.y = mirrorbase + mMirrorOffset.y;
     int diff = newpos.y-oldpos.y;
     newpos.y += diff;
