+++ Adventure/AppBuilder/AppBuilder/Engine.h
@@ -50,7 +50,7 @@ public:
   void unloadRoom(RoomObject* room, bool mainroom, bool immedately);
   bool setFocus(std::string charname, ExecutionContext* reason);
   CGE::Image* getImage(const std::string& name);
-  int getWalkGridSize(bool doublewalkmap) {return doublewalkmap ? mWalkGridSize/2 : mWalkGridSize;}
+  float getWalkGridSize(bool doublewalkmap) {return doublewalkmap ? mWalkGridSize/2.0f : (float)mWalkGridSize;}
   Object2D* getObjectAt(const Vec2i& pos);
   Object2D* getObject(const std::string& name, bool searchInventoryFirst);
   CharacterObject* getCharacter(const std::string& name);
