+++ Adventure/AppBuilder/AppBuilder/EngineObjects.h
@@ -192,7 +192,7 @@ public:
   void setDepth(int depth);
   void setFadeout(int time) {mFadeout = time;}
   int getFadeout() {return mFadeout;}
-  int getWalkGridSize();
+  float getWalkGridSize();
 protected:
   class DepthMap {
   public:
@@ -268,7 +268,7 @@ public:
   void updateState(bool mirror, bool force);
   //virtual bool animationEnded(Animation* anim);
   virtual void activateNextState();
-  int getWalkGridSize();
+  float getWalkGridSize();
 protected:
   static int calculateState(LookDir dir, bool& mirror);
   static int calculateState(int currState, bool shouldWalk, bool shouldTalk, bool mirror=false);
